+++ templates/Agterhuis.Ui.Templates/TemplatePackage.cs
@@ -1,0 +1,5 @@
+namespace Agterhuis.Ui.Templates;
+
+internal static class TemplatePackage
+{
+}